Revert "ui_test done"
This reverts commit b05a33e0d70aa6d86b9a2b575abd8539142d4c0d.
