+++ app/build.gradle
@@ -68,9 +68,9 @@ dependencies {
 
     // JUnit
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
+    testImplementation 'junit:junit:4.12'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.3.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.3.2'
 
     // Roboelectric
     androidTestImplementation "androidx.test:runner:1.5.2"
